+++ Frontend/src/app/features/formats/components/row-area/row-area.component.html
