+++ Frontend/src/app/features/formats/components/row-single/row-single.component.html
