+++ Frontend/src/app/features/formats/components/row-single/row-single.component.ts
@@ -1,14 +1,14 @@
 import { ChangeDetectionStrategy, Component, inject, input, Input, OnInit, output } from '@angular/core';
 import { FormGroup, NonNullableFormBuilder, Validators } from '@angular/forms';
-import { RowsFormType } from '@features/format/format.component';
 import { MaterialModule } from '@shared/modules/material/material.module';
 import { INPUT_TYPES, InputTypes } from '@tipos/format';
-import { InputTextComponent } from '../input-text/input-text.component';
 import { CdkDrag, CdkDragHandle } from '@angular/cdk/drag-drop';
+import { InputTextComponent } from '../input-text/input-text.component';
 import { InputSelectComponent } from '../input-select/input-select.component';
+import { RowsFormType } from '../format/format.component';
 
 @Component({
-  selector: 'app-row-table',
+  selector: 'app-row-single',
   standalone: true,
   imports: [
     MaterialModule,
@@ -16,11 +16,11 @@ import { InputSelectComponent } from '../input-select/input-select.component';
     InputTextComponent,
     InputSelectComponent,
   ],
-  templateUrl: './row-table.component.html',
-  styleUrl: './row-table.component.scss',
+  templateUrl: './row-single.component.html',
+  styleUrl: './row-single.component.scss',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class RowTableComponent implements OnInit {
+export class RowSingleComponent implements OnInit {
   private formBuilder = inject(NonNullableFormBuilder);
 
   @Input({ required: true }) row!: FormGroup<RowsFormType>;
@@ -34,7 +34,7 @@ export class RowTableComponent implements OnInit {
     { type: INPUT_TYPES.DATE, name: 'Fecha' },
   ];
 
-  ngOnInit(): void {
+  ngOnInit() {
     if (!this.row.controls.fields.length) this.addColumn(INPUT_TYPES.TEXT);
   }
 
@@ -53,4 +53,5 @@ export class RowTableComponent implements OnInit {
   removeRow(): void {
     this.remove.emit();
   }
+
 }
